Create Get-RSSFeed.ps1
Write a short script that can retrieve the most recent article headlines from a blog by using the blog’s RSS or Atom feed. You should ideally just display the headlines, but might also choose to display a URL that links to the article, and might display a short excerpt of the article. If the feed contains the full article text, don’t display it – at most, display a short excerpt.

While you could definitely write this as a one-liner, and might choose to do so as you start, there’s real value in turning this into a “Get-RSSFeed” function. To be fair, lots of folks have done this before – but challenge yourself, and try to figure it out without opening a search engine!

 

Challenges:

Try to write this to be a PowerShell command (an advanced function) that uses parameters to direct the behaviour of the command.
Try to ensure your script’s output could be easily displayed in an on-screen table, or redirected to a CSV file.
Try to minimize your use of “raw” .NET classes (e.g., try to use only PowerShell commands as much as possible).